+++ Common/src/main/java/fuzs/moblassos/capability/VillagerContractCapability.java
@@ -3,8 +3,9 @@ package fuzs.moblassos.capability;
 import fuzs.moblassos.MobLassos;
 import fuzs.moblassos.config.ServerConfig;
 import fuzs.puzzleslib.api.capability.v2.data.CapabilityComponent;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.npc.AbstractVillager;
+import net.minecraft.world.entity.npc.Villager;
+import net.minecraft.world.entity.npc.VillagerData;
 
 public interface VillagerContractCapability extends CapabilityComponent {
 
@@ -12,13 +13,12 @@ public interface VillagerContractCapability extends CapabilityComponent {
 
     boolean hasAcceptedContract();
 
-    static boolean canAcceptContract(Entity entity) {
-        double villagerAcceptsContractChance = MobLassos.CONFIG.get(ServerConfig.class).villagerAcceptsContractChance;
-        if (villagerAcceptsContractChance == 0.0) return false;
-        if (entity instanceof LivingEntity livingEntity && livingEntity.isBaby()) {
-            villagerAcceptsContractChance = Math.min(1.0, villagerAcceptsContractChance * 2.0);
+    static boolean canAcceptContract(AbstractVillager abstractVillager) {
+        if (!MobLassos.CONFIG.get(ServerConfig.class).villagerAcceptsContract) return false;
+        if (abstractVillager instanceof Villager villager) {
+            return Math.abs(villager.getUUID().getLeastSignificantBits() % VillagerData.MAX_VILLAGER_LEVEL) < villager.getVillagerData().getLevel();
+        } else {
+            return true;
         }
-        int acceptsChance = (int) (1.0 / villagerAcceptsContractChance);
-        return entity.getUUID().getLeastSignificantBits() % acceptsChance == 0;
     }
 }