+++ Common/src/main/java/fuzs/moblassos/config/ServerConfig.java
@@ -28,9 +28,8 @@ public class ServerConfig implements ConfigCore {
     @Config(description = {"Time interval in seconds after which the player is hurt by half a heart for carrying any hostile lasso containing a monster.", "Set to -1 to disable."})
     @Config.IntRange(min = -1)
     public int hostileDamageRate = 5;
-    @Config(description = "The chance for a villager to accept a contract allowing it to be captured by an emerald lasso.")
-    @Config.DoubleRange(min = 0.0, max = 1.0)
-    public double villagerAcceptsContractChance = 0.2;
+    @Config(description = "Can a villager accept a contract allowing it to be captured by an emerald lasso.")
+    public boolean villagerAcceptsContract = true;
     @Config(description = "The percentage each level of the holding enchantment will increase a lasso's holding time by.")
     @Config.DoubleRange(min = 0.0)
     public double holdingMultiplier = 0.2;