+++ CHANGELOG.md
@@ -3,7 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v6.0.0-1.19.4] - 2023-05-25
-- Ported to Minecraft 1.19.4
+## [v7.0.0-1.20] - 2023-06-10
+- Ported to Minecraft 1.20
+### Changed
+- All villagers can now accept a contract, this now depends on the level (novice, journeyman, master, ...) of the villager
+- Each level will increase the chance for accepting a contract by 20%, so a master level villager is guaranteed to accept a contract if no lower level worked 
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/